Uncommented other tests, linting errors resolved, refactoring directory structure, using a randomized key for pallet_storage_test now.
