+++ tools/two_address_substrate_with_extrinsic.py
@@ -372,55 +372,6 @@ def pallet_batchall_test():
         print("⚠️ No local Substrate node running, try running 'start_local_substrate_node.sh' first")
         sys.exit()
 
-def storage_add_item(substrate, kp_src, item_type, item):
-    nonce = substrate.get_account_nonce(kp_src.ss58_address)
-    call = substrate.compose_call(
-        call_module='PeaqStorage',
-        call_function='add_item',
-        call_params={
-            'item_type': item_type,
-            'item': item,
-        })
-
-    extrinsic = substrate.create_signed_extrinsic(
-        call=call,
-        keypair=kp_src,
-        era={'period': 64},
-        nonce=nonce
-    )
-
-    receipt = substrate.submit_extrinsic(extrinsic, wait_for_inclusion=True)
-    show_extrinsic(receipt, 'add_item')
-
-    if not receipt.is_success:
-        print(substrate.get_events(receipt.block_hash))
-        raise IOError
-
-def utf8_to_ascii(utf8str):
-    return [int(utf8str[i:i+2],16) for i in range(0,len(utf8str),2)]
-
-def storage_rpc_read(substrate, kp_src, item_type, item):
-    data = substrate.rpc_request('peaqstorage_readAttribute', [kp_src.ss58_address, item_type])
-    # TODO RPC returns an array of bytes, not key-value struct as is expected.
-    assert(binascii.hexlify(bytes(data["result"])) == bytes(item, 'utf-8')[2:])
-
-def pallet_storage_test():
-    print('---- pallet_storage_test!! ----')
-    try:
-        # Check the type_registry_preset_dict = load_type_registry_preset(type_registry_name)
-        # ~/venv.substrate/lib/python3.6/site-packages/substrateinterface/base.py
-        with SubstrateInterface(url=WS_URL) as substrate:
-            kp_src = Keypair.create_from_uri('//Alice') 
-            item_type = utf8_to_ascii("abcc")
-            item = '0x032132'
-
-            storage_add_item(substrate, kp_src, item_type, item)
-            storage_rpc_read(substrate, kp_src, item_type, item)
-
-    except ConnectionRefusedError:
-        print("⚠️ No local Substrate node running, try running 'start_local_substrate_node.sh' first")
-        sys.exit()
-
 if __name__ == '__main__':
     # kp_src = Keypair.create_from_mnemonic('nature exchange gasp toy result bacon coin broccoli rule oyster believe lyrics')
     # print(kp_src.ss58_address)
